update test to address dplyr 1.0.0
I'm not sure why it's happening now and not before, but it was an issue
of the non-df form returning a data frame and the df form returning a
tibble. I tried to convert the tibble to a data frame if that was what
was entered, but that screwed up a lot of other tests. This is NOT the
best way to do this, but I really don't want to waste time here.